@@ -1,33 +1,33 @@
 {
   "name": "chewsr",
-  "version": "0.1.0",
+  "version": "2.0.0",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-free": "^5.15.4",
-    "@fortawesome/fontawesome-svg-core": "^1.2.36",
+    "@fortawesome/fontawesome-free": "^6.5.1",
+    "@fortawesome/fontawesome-svg-core": "^6.5.1",
     "@fortawesome/free-regular-svg-icons": "^6.5.1",
-    "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@fortawesome/react-fontawesome": "^0.1.19",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
-    "bootstrap": "^4.6.2",
-    "firebase": "^7.24.0",
-    "firebase-admin": "^8.13.0",
-    "firebase-tools": "^7.16.2",
-    "google-map-react": "^1.1.7",
+    "@fortawesome/free-solid-svg-icons": "^6.5.1",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "@testing-library/jest-dom": "^6.2.0",
+    "@testing-library/react": "^14.1.2",
+    "@testing-library/user-event": "^14.5.2",
+    "bootstrap": "^5.3.2",
+    "firebase": "^10.7.2",
+    "firebase-admin": "^12.0.0",
+    "firebase-tools": "^13.0.3",
+    "google-map-react": "^2.2.1",
     "hammerjs": "^2.0.8",
     "jquery": "^3.7.1",
-    "node-sass": "^4.14.1",
-    "rc-slider": "^9.7.5",
-    "react": "^16.14.0",
-    "react-bootstrap": "^1.6.8",
-    "react-cookie": "^4.1.1",
-    "react-dom": "^16.14.0",
+    "node-sass": "^9.0.0",
+    "rc-slider": "^10.5.0",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.10.0",
+    "react-cookie": "^7.0.1",
+    "react-dom": "^18.2.0",
     "react-google-maps-loader": "^4.3.1",
     "react-google-places-suggest": "^3.12.3",
-    "react-scripts": "^3.4.4",
-    "style-loader": "^3.3.3"
+    "react-scripts": "^5.0.1",
+    "style-loader": "^3.3.4"
   },
   "scripts": {
     "start": "react-scripts start",
